requirements.txt
paddlepaddle
paddleocr
pdf2image
Pillow
openpyxl
requests
python-dotenv
PyMuPDF
